requirements.txt
numpy==1.15.3
Pillow==5.3.0
scipy==1.1.0
tensorflow==1.11.0
h5py==2.8.0
tqdm==4.28.1
